Move Audio,SoundPlayer out from Music Part
